+++ src/main/Main.hs
@@ -537,12 +537,20 @@ pathCmd keys go =
             distDir <- distRelativeDir
             hpcDir <- hpcReportDir
             compilerPath <- getCompilerPath =<< getWhichCompiler
+            when (deprecatedStackRootOptionName `elem` keys) $
+                liftIO $ forM_
+                    [ ""
+                    , "'--" <> deprecatedStackRootOptionName <> "' will be removed in a future release."
+                    , "Please use '--" <> stackRootOptionName <> "' instead."
+                    , ""
+                    ]
+                    (T.hPutStrLn stderr)
             forM_
                 -- filter the chosen paths in flags (keys),
                 -- or show all of them if no specific paths chosen.
                 (filter
                      (\(_,key,_) ->
-                           null keys || elem key keys)
+                           (null keys && key /= deprecatedStackRootOptionName) || elem key keys)
                      paths)
                 (\(_,key,path) ->
                       liftIO $ T.putStrLn
@@ -583,7 +591,7 @@ data PathInfo = PathInfo
 -- | The paths of interest to a user. The first tuple string is used
 -- for a description that the optparse flag uses, and the second
 -- string as a machine-readable key and also for @--foo@ flags. The user
--- can choose a specific path to list like @--global-stack-root@. But
+-- can choose a specific path to list like @--stack-root@. But
 -- really it's mainly for the documentation aspect.
 --
 -- When printing output we generate @PathInfo@ and pass it to the
@@ -592,7 +600,7 @@ data PathInfo = PathInfo
 paths :: [(String, Text, PathInfo -> Text)]
 paths =
     [ ( "Global stack root directory"
-      , "global-stack-root"
+      , stackRootOptionName
       , T.pack . toFilePathNoTrailingSep . configStackRoot . bcConfig . piBuildConfig )
     , ( "Project root (derived from stack.yaml file)"
       , "project-root"
@@ -647,7 +655,20 @@ paths =
       , T.pack . toFilePathNoTrailingSep . piDistDir )
     , ( "Where HPC reports and tix files are stored"
       , "local-hpc-root"
-      , T.pack . toFilePathNoTrailingSep . piHpcDir ) ]
+      , T.pack . toFilePathNoTrailingSep . piHpcDir )
+    , ( "DEPRECATED: Use '--" <> T.unpack stackRootOptionName <> "' instead"
+      , deprecatedStackRootOptionName
+      , T.pack . toFilePathNoTrailingSep . configStackRoot . bcConfig . piBuildConfig )
+    ]
+
+stackRootOptionName :: Text
+stackRootOptionName = "stack-root"
+
+-- Deprecated since stack-1.0.5.
+-- TODO: Remove occurences of this variable and use 'stackRootOptionName' only
+-- after an appropriate deprecation period.
+deprecatedStackRootOptionName :: Text
+deprecatedStackRootOptionName = "global-stack-root"
 
 data SetupCmdOpts = SetupCmdOpts
     { scoCompilerVersion :: !(Maybe CompilerVersion)